Separate definition and declaration of utility tools
