+++ lib/CMakeLists.txt
@@ -2,6 +2,7 @@ add_subdirectory(Dialect)
 add_subdirectory(Conversion)
 add_subdirectory(Target)
 add_subdirectory(Interface)
+add_subdirectory(Utils)
 
 add_mlir_library(static_mlir_async_runtime
   STATIC
