+++ lib/Utils/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(DIPUtils DIPUtils.cpp)
+add_library(Utils Utils.cpp)
